+++ app/routers/geongan/fcompany.py
@@ -3,31 +3,31 @@ from sqlalchemy.orm import Session
 
 from app.core.database import get_db
 from app.core.security import get_current_user
-from app.repository.geogan.fdept_repo import create_fdept, get_all_fdept
-from app.schemas.geogan.fdept import FDeptCreate, FDeptResponse
-from app.models.geogan.fdept import FDept
+from app.repository.geongan.fcompany_repo import create_fcompany, get_all_fcompany
+from app.schemas.geongan.fcompany import FCompanyCreate, FCompanyResponse
+from app.models.geongan.fcompany import FCompany
 
-router = APIRouter(prefix="/api/geogan", tags=["fdept"])
+router = APIRouter(prefix="/api/geongan", tags=["fcompany"])
 
-@router.post("/createFDept", response_model=FDeptResponse)
-def create_fdept_endpoint(
-    payload: FDeptCreate,
+@router.post("/createFCompany", response_model=FCompanyResponse)
+def create_fcompany_endpoint(
+    payload: FCompanyCreate,
     db: Session = Depends(get_db),
     current_user = Depends(get_current_user),
 ):
     if "ROLE_ADMIN" not in current_user["roles"]:
         raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Admin only")
-    existing = db.query(FDept).filter_by(id=payload.id).first()
+    existing = db.query(FCompany).filter_by(id=payload.id).first()
     if existing:
         raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="ID already exists")
-    result = create_fdept(db, payload)
-    return FDeptResponse.model_validate(result)
+    result = create_fcompany(db, payload)
+    return FCompanyResponse.model_validate(result)
 
-@router.get("/getAllFDept", response_model=list[FDeptResponse])
-def read_all_fdept(
+@router.get("/getAllFCompany", response_model=list[FCompanyResponse])
+def read_all_fcompany(
     db: Session = Depends(get_db),
     current_user = Depends(get_current_user),
 ):
     if "ROLE_ADMIN" not in current_user["roles"]:
         raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Admin only")
-    return get_all_fdept(db)
+    return get_all_fcompany(db)
