+++ app/routers/geongan/fdatabase.py
@@ -3,31 +3,31 @@ from sqlalchemy.orm import Session
 
 from app.core.database import get_db
 from app.core.security import get_current_user
-from app.repository.geogan.fcompany_repo import create_fcompany, get_all_fcompany
-from app.schemas.geogan.fcompany import FCompanyCreate, FCompanyResponse
-from app.models.geogan.fcompany import FCompany
+from app.repository.geongan.fdatabase_repo import create_fdatabase, get_all_fdatabase
+from app.schemas.geongan.fdatabase import FDatabaseCreate, FDatabaseResponse
+from app.models.geongan.fdatabase import FDatabase
 
-router = APIRouter(prefix="/api/geogan", tags=["fcompany"])
+router = APIRouter(prefix="/api/geongan", tags=["fdatabase"])
 
-@router.post("/createFCompany", response_model=FCompanyResponse)
-def create_fcompany_endpoint(
-    payload: FCompanyCreate,
+@router.post("/createFDatabase", response_model=FDatabaseResponse)
+def create_fdatabase_endpoint(
+    payload: FDatabaseCreate,
     db: Session = Depends(get_db),
     current_user = Depends(get_current_user),
 ):
     if "ROLE_ADMIN" not in current_user["roles"]:
         raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Admin only")
-    existing = db.query(FCompany).filter_by(id=payload.id).first()
+    existing = db.query(FDatabase).filter_by(id=payload.id).first()
     if existing:
         raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="ID already exists")
-    result = create_fcompany(db, payload)
-    return FCompanyResponse.model_validate(result)
+    result = create_fdatabase(db, payload)
+    return FDatabaseResponse.model_validate(result)
 
-@router.get("/getAllFCompany", response_model=list[FCompanyResponse])
-def read_all_fcompany(
+@router.get("/getAllFDatabase", response_model=list[FDatabaseResponse])
+def read_all_fdatabase(
     db: Session = Depends(get_db),
     current_user = Depends(get_current_user),
 ):
     if "ROLE_ADMIN" not in current_user["roles"]:
         raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Admin only")
-    return get_all_fcompany(db)
+    return get_all_fdatabase(db)
