+++ test/geongan/test_fcompany.py
@@ -2,5 +2,5 @@ import importlib
 
 class TestFCompany:
     def test_import_router(self):
-        module = importlib.import_module("app.routers.geogan.fcompany")
+        module = importlib.import_module("app.routers.geongan.fcompany")
         assert hasattr(module, "router")
