+++ test/geongan/test_fdatabase.py
@@ -2,5 +2,5 @@ import importlib
 
 class TestFDatabase:
     def test_import_router(self):
-        module = importlib.import_module("app.routers.geogan.fdatabase")
+        module = importlib.import_module("app.routers.geongan.fdatabase")
         assert hasattr(module, "router")
