+++ test/geongan/test_fdept.py
@@ -2,5 +2,5 @@ import importlib
 
 class TestFDept:
     def test_import_router(self):
-        module = importlib.import_module("app.routers.geogan.fdept")
+        module = importlib.import_module("app.routers.geongan.fdept")
         assert hasattr(module, "router")
