+++ test/geongan/test_fdivision.py
@@ -2,5 +2,5 @@ import importlib
 
 class TestFDivision:
     def test_import_router(self):
-        module = importlib.import_module("app.routers.geogan.fdivision")
+        module = importlib.import_module("app.routers.geongan.fdivision")
         assert hasattr(module, "router")
