+++ test/geongan/test_ffiles_lainnya.py
@@ -2,5 +2,5 @@ import importlib
 
 class TestFFilesLainnya:
     def test_import_router(self):
-        module = importlib.import_module("app.routers.geogan.ffiles_lainnya")
+        module = importlib.import_module("app.routers.geongan.ffiles_lainnya")
         assert hasattr(module, "router")
