+++ test/geongan/test_ffiles_mutasi.py
@@ -2,5 +2,5 @@ import importlib
 
 class TestFFilesMutasi:
     def test_import_router(self):
-        module = importlib.import_module("app.routers.geogan.ffiles_mutasi")
+        module = importlib.import_module("app.routers.geongan.ffiles_mutasi")
         assert hasattr(module, "router")
