+++ test/geongan/test_fproses_mutasi.py
@@ -2,5 +2,5 @@ import importlib
 
 class TestFProsesMutasi:
     def test_import_router(self):
-        module = importlib.import_module("app.routers.geogan.fproses_mutasi")
+        module = importlib.import_module("app.routers.geongan.fproses_mutasi")
         assert hasattr(module, "router")
