+++ test/geongan/test_fstatus_mutasi.py
@@ -2,5 +2,5 @@ import importlib
 
 class TestFStatusMutasi:
     def test_import_router(self):
-        module = importlib.import_module("app.routers.geogan.fstatus_mutasi")
+        module = importlib.import_module("app.routers.geongan.fstatus_mutasi")
         assert hasattr(module, "router")
